WL#8069 Fixed bug in DIH where we assumed that no LCP could have been saved and still no LCP valid found. With 8069 we report LCP completed much later, so this means that this is a perfectly legitimate event.
We kept around some checks that we didn't start an LCP with a queue.
This kept around check was buggy, so removed it and replaced it with
a simple check at start of LCP that queue is empty.

Fixed one bug previously with handling of skipped pages, however this
meant that the variable all_parts got to be always false and we entered
into the Partial LCP code. The partial LCP code needed to initialise the
m_tableId in the scan record, but we also needed to always set the
all_parts to true to ensure that we don't use the Partial LCP code just
yet. Will require some fixing later on when adding Partial LCP for real.