wl#10234
Enhance TC block to handle Multi fragment scan requests:

A Multi-Fragment scan request will try to 'group' as many
fragment scan request to the same 'node/instance' as possible
into the same GSN_SCAN_FRAGREQ signal.

In order to facilitate this we have to know which LDM instance
each fragment is located on before we produce any GSN_SCAN_FRAGREQ
signals.

Thus, the Dbtc code had to be enhanced such that instead of
immediately sending SCAN_FRAGREQ when DIH had replied on
a DIGETNODESREQ-signal, we now instead collect this info into
the fragLocations[]-array. When this array has been filled in
for all fragments, it is sorted on the fragments 'blockRef'.
(Which is the blockRef of the SPJ block to execute on in case
of a MultiFrag-viaSPJ scan request)

We are then ready to enter the new sendFragScansLab phase which
use the collected fragLocations to send SCAN_FRAGREQ's to
LQH / SPJ.

Due to splitting the old combined DIGETNODESREQ + SCAN_FRAGREQ
phase into two seperate phases some more restructuring
of the code, mostly related to error handling:

- Several places we could remove code to 'release' a failed
  scanFrag from the 'm_running_scan_frags' list. (Due to
  we were still being in the 'collect fragLocations[]' phase
  where we could not be 'running' yet)

- Code for checking 'CLOSING_SCAN' and 'scanSchemaVersion'
  had to be adapted to the new phases.