wl#10234 Rename patch
Renames class NdbRootFragment to class NdbWorker and update related
comments to reflect this change.

Class NdbWorker (former NdbRootFragment) handles the results
being produced from a single SPJ SCANFRAG request, starting by
scanning a specific root fragment. A full SPJ execution required
one SPJ-SCANFRAG request to be sent to a SPJ instance for each
root fragment to scan.

With this 'multi-fragment' worklog that will change such that
a SPJ SCANFRAG request may scan multiple root fragments. Thus
the term NdbRoot*Fragment* was a bit misleading.

The intention of the new naming is to focus on that execution
of a SPJ query employes multiple SPJ instances as 'workers' - each
of them handling some of the fragments.