wl#10234: Enable usage of MultiFrag scan from API -> TC -> SPJ
SPJ API sets the ScanTabReq::setMultiFragFlag() to enable the feature.
'm_fragsPrWorker' is calculated by the API such that each 'worker'
handle all fragments located on the same data node.

As a 'batch-buffer is set up pr fragment, the batch size available
for each combined MultiFrag scan is multiplies by m_fragsPrWorker.

Dbspj::execSCAN_FRAGREQ receives the list of fragments to scan
(set up by TC) in the last longSignal section. Instead of
scanFrag_build() setting up a single ScanFragHandle for the
root scan, we now construct one for each of the MultiFragment.

As there is an implementation defined 12-bit limit on
'maxBatchRows', some additional checks had to be introduced
to avoid it to overflow. (Due to larger batch sizes now being
available)