Bug #26023259: REMOVE RADIXSORT
We have special-purpose radixsort code for sorting very short keys, but in
practical testing, std::stable_sort is now faster. Thus, we can just remove the
radixsort.

Verified by sorting 50000 short records on an int field, which is supposed to
be the poster child for radxisort; removing the code was 10–20% faster in
single-threaded testing, and there should be yet more in multithreaded, due to
less cache pressure.

Change-Id: I4f17344e7a744ee56df42cabd9423d9ccd93ba9c