Bug #26030132: REMOVE SQL_ALLOC
Make all allocation on THR_ALLOC explicit. This is a no-op change;
even nonsensical allocations on THR_ALLOC (ie., where we obviously
have a THD as a local variable that we could use, instead of pulling it
out of TLS) are kept. This exposes all use of the alloc stored in TLS
to make it explicit, and removes the difference in syntax between
Sql_alloc-derived classes and all others, increasing consistency.

We remove the replacement operator new on Sql_alloc that used to use
THR_ALLOC behind-the-scenes; this means it will be impossible to write
new Foo() without an explicit MEM_ROOT by mistake (since overload resolution
will fail). We'll keep it this way until 8.0 has been released, when we can
remove Sql_alloc entirely and new Foo() will start meaning simply to allocate
Foo() the regular way.

A few select places still depend on operator delete, which is kept
until the next patch.

Change-Id: I9f1af8762eb3bee79cc0a2492f23289a1f4b78d6