Bug #26030132: REMOVE SQL_ALLOC
Make all destruction-but-not-deletion of Sql_alloc objects explicit.
This means that it's now possible to see (without looking at the object's
type and finding the reference to Sql_alloc in the class hierarchy)
that "delete" of such objects does not actually free any memory.

To enforce this, we make operator delete in Sql_alloc protected
(it can't be private, or objects could not be destructed at all),
making every such delete call fail.

Change-Id: I7fdf6dc6099f1ecfcec9131bae6b6dd2d77b3ec1