Bug#26040934 REMOVE OBSOLETE INCLUDE_DIRECTORIES IN CMAKE FILES
This is a followup to the patch for
Bug #25436469 BUILDS ARE NOT REPRODUCIBLE

perl will sometimes report
Use of uninitialized value $dir1 in pattern match (m//) at
scripts/invoke-with-relative-paths.pl line 62.

This turns out to be obsolete/wrong cmake INCLUDE_DIRECTORIES directives.

Change-Id: I9d63ff273720ddc150880f9d4f251473b7f9e1df
Fix: print a warning in the perl script if an unknown path is detected.
Also: remove all obsolete INCLUDE_DIRECTORIES from .cmake files.