Bug#26042934: Unchecked read after allocated buffer
Problem:
There is no check for packet length before reading the byte that
specifies if there are new data types provided.

Solution:
Added checks for the length of the packet_left before reading the byte.
Also added an assert and a check for the packet length left while
reading each ps parameter.

Change-Id: I059c7eb31cd5005f4bc8aee3048fdcc2221ff381