Bug #24829435 	NDB : NDBTABLE::GETCOLUMN(CONST CHAR*) IS INEFFICIENT
The getColumn functionality in the NdbApi dictionary is used to
map from a column name to a column object.
This can be used explicitly by users, or implicitly as part of
passing a column name to an NdbApi function.
The current implementation uses a linear search of an array of
strings (column names), whose cost scales O(n) in the number of
columns.  This has been seen to waste cpu in real-world NdbApi
applications.
It is possible for applications to lookup column objects or ids
upfront, and then use them which gives O(1) access to a column
object.
However to improve the experience of users using column name strings
at runtime, a column name hash is implemented.  This should give
~O(1) lookup performance from name strings to column objects.

It is still recommended to perform name->object/id lookups upfront,
but the cost of doing them at runtime is reduced.

A new test is added to testNdbApi which can be used to show
that column lookup performance is independent of the #columns.