Bug#25858461: FAILURE TO DROP DATAFILE FROM TABLESPACES WITH MULTIPLE DATAFILES
Problem: It was not possible to drop a tablespace file from a general
tablespace which had more than one tablespace file.

Root cause was that when a tablespace file is dropped, all tablespace
files with a higher ordinal position will have their ordinal position
decremented by one. Since the ordinal position is part of the PK of
TABLESPACE_FILES, the store following the drop would result in an
pk-update of the files which had their ordinal position decremented.

But pk-updates are not supported by the store-implementation in
dd::Weak_object_impl as it searches for the new key and only performs
an update if this key is found. When the key itself is being updated
searching for the new key is wrong and could cause the wrong tuple to
be updated, or as in this case, a new tuple to be inserted without
deleting the tuple with the old key. In this case such an insert was
rejected because both the old and the new tuple had the same value for
TABLESPACE_FILES.file_name, and there is a unique constraint on this
column.

Solution: When dropping a tablespace file, all the files must be
dropped from TABLESPACE_FILES first, and then the remaining files (that
were not supposed to be dropped) must be re-insterted with the correct
PK and file_name, so that duplicates are avoided.

(cherry picked from commit a6de211c736f49a74ad433f7651b629552c9115d)