Bug#25804386 - DUPLICATE VALUE FOR AUTOINC COLUMN
Background:
----------
On checkpoint, the DDTableBuffer in-memory contents are written to the
physical DDBuffer Table.

During checkpoint, the redo generated by these DDTableBuffer writes
are written to disk.

Before WL#9499:
--------------
The fil_name_parse() which writes MLOG_FIL_NAMEs made sure that the
redo generated by DDBuffer Table writes are also flushed to disk

So this is not so obvious dependency on fil_names_clear().

After WL#9499:
--------------
Since MLOG_FILE_NAME is removed, there is no need to write MLOG_FILE_NAMEs
as part of checkpoint. So fil_names_clear() is removed.

This exposed the above explained dependency of DDTableBuffer writes
on fil_names_clear(). Since the redo-writes of DDTableBuffer didn't
make to disk on checkpoint, after a crash, a duplicate auto-inc value
is possible and hence this bug.

Fix:
----
1. Change return type of dict_persist_to_dd_table_buffer() to determine if a possible
   write to DDTableBuffer happenned.

2. Flush the redo upto a lsn which includes DDTableBuffer changes.

Reviewed-By: Debarun Banerjee <debarun.banerjee@oracle.com>
Reviewed-By: Bin Su <bin.x.su@oracle.com>
RB: 16071