BUG#26020990 --PRINT-TABLE-METADATA DOES NOT DISPLAY ALL METADATA IN TABLE_MAP_LOG EVENT
Problem:
WL#4618 introduced additional metadata to Table_map_log_event and the
option to have it printed by mysqlbinlog. However, when using this
option, some metadata already stored in this event was not printed.

Fix:
When a column is marked as NOT NULL, either explicitly or implicitly
(for instance, when it is a Primary Key), NOT NULL is printed after its
type.
Fixed typo INVALIDE_GEOMETRY_TYPE.
Fixed spacing when the column name is not printed and when the types SET
and ENUM are printed.