Bug #26089518: SPORADIC FAILURE OF INNODB.INNODB-ISOLATION
Problem
-------
The results diff is because the explain output
of the query sometimes gives the number of rows
slightly different that what is stored in result file.
This is because the row count variable provided by
the explain is merely a estimate and  can vary.

FIX
---

As done in other parts of test case we have
we are ignoring the number of rows returned
by the explain query. We make sure that plans
are not changed by using analyze.

[ #rb 16267 Pushing on behalf  Sree Harsha ]