wl#10234, Reintroduce building of SPJ performance measurement tool
spj_performance_test.cpp was one of the tools originaly used
when developing the SPJ feature. It turns out that building of this
tool was disabled when we switched to using cmake.

Patch reintroduce build of this tool and also extend it to
be able to test Scan-Scan spj queries. (Spj queries where
both the root node and its childs are scans)