Bug#26113652 - DON'T INHERIT TABLESPACE WHEN DATA DIRECTORY IS SPECIFIED FOR PARTITIONED TABLE
The problem is that if a partition is specified without TABLESPACE
but with DATA DIRECTORY, it may report TABLESPACE and DATA DIRECTORY
conflicttion if there is higher level TABELSPACE specified.

The solution should be that if DATA DIRECTORY is specified for a partition,
it should always ignore higher level TABLESPACE as how we do before.

Reviewed-by: Jimmy Yang<Jimmy.Yang@oracle.com>
RB: 16305