BUG#26078602: BACKPORT BUG#25574200 TO 5.7
Issue:
======
For a given below scenario where S1 and S2 are in Group Replication in Single
Primary Mode and S3 and S2 are in asynchronous replication. In Group Replication
between S1 and S2, S1 acts as Primary and S2 as secondary. In Asynchronous
Replication between S3 and S2, S3 acts as master and S2 as slave.

   --Single Primary Mode--
   |                     |
Primary              Secondary
  S1<------------------->S2<--------------------S3
      Group Replication      Async Replication

Now issue comes when any data is inserted in S3 which gets replicated to
S2 which has super_read_only mode active due to single primary mode. But
super_read_only mode only prohibits client updated and so it can get data
from master. And as S1 and S2 are in Group Replication data gets through
to S1 from a read-only member.

Resolution:
===========
Don't allow asynchronous channel to start on secondary when group replication
is in single primary-mode.
Don't allow group replication to start on single primary-mode when
asynchronous replication channels are running.

Note:

This is a backport of Bug#25574200 patch from mysql-8.0.