Bug#25923125 - 7.6.2: NODE CRASHED DURING RESTART WITH ERROR "COPYFRAG FAILED, ERROR: 1601
If the tablespace of a disk table is fully consumed when a node goes
down and the table rows are deleted and inserted while the node is
away (or updated with shrinking/expanding disk column values), the
node restart may fail with error 1601 - "Out of extents, tablespace
full".

This is fixed by reserving 4% of the tablespace to be used during
node starts.