Bug#25907510 CRASH IN HA_MYISAM::IDX_COND_PUSH WITH INTERNAL_TMP_DISK_STORAGE_ENGINE=MYISAM
In WL#883 (CTEs) this was changed, for a tmp table created for a
derived table:
- before: one such table has a single TABLE;
TABLE::key_info and TABLE::s::key_info are the same pointer,
so accessing the array of key descriptions is possible from any of the
two pointers.
- after: multiple TABLEs are possible for CTEs referenced more than once,
so pointers are different pointers, as some keys can exist in a
TABLE_SHARE while not yet existing in all TABLEs connected to it
(if in the middle of the key creation process).
Thus, content of key_info for the TABLE and TABLE_SHARE can be different.
Content for TABLE_SHARE's is a shallow version of that of TABLE's,
filled in TABLE::add_tmp_key().
For example, key_info::key_part_info isn't filled for TABLE_SHARE.

Alas I overlooked that some old code, ha_myisam::idx_cond_push(),
uses TABLE_SHARE::key_info and has a valid reason to do so:
- it has no access to TABLE (see fix for Bug 12976163).
- what it's reading from TABLE_SHARE::key_info is identical for
all TABLEs of this share, so the read is safe.

Fix: in add_tmp_key(), delay the copying by a few lines, so that
the TABLE's key_info is more complete when the copying happens,
so that the TABLE_SHARE's key_info is more complete too.