BUG#26042764: BACKPORT BUG#25642343 AND BUG#24311527 TO 5.7 BUG#24311527: PLUGIN-LOAD FAILS AFTER MYSQLD --INITIALIZE BUG#25642343: THE SERVER SHOULD HAVE A DEFAULT USER FOR THE SERVER SESION SERVICE
Backport of BUG#25642343 AND BUG#24311527 into mysql-5.7

Plugins like X and Group Replication try to use the root user to
contact the server using the Server Session Service.
This approach has many drawbacks as the root user may not be in the
server or is expired.
This patch proposes a different approach by adding to the server a
default locked user that can be used by plugins.

This patch introduces:
  - The user creation on --initialize
  - The user creation on upgrade
  - MTR changes so the user is created on bootstrap

This patch also includes changes to the way the group replication
plugin contacts the server to protect the context associated to
external users sessions.