Bug#26084820: CHARACTER SET PROBLEMS IN INNODB_FTS.PROXIMITY TEST
Issue:
------
The innodb_fts.proximity contains Unicode characters, but does not
set the character set to uft8, so by default on 5.7, latin1 is used.
This means that the Chinese characters in the test are not
interpreted as Unicode characters, but as latin1 characters. They
are then converted to utf8 before being inserted into the table
which is created with utf8 character set. This results is some queries
returning an incorrect number of rows because certain characters are not
matched.

Fix:
----
Set the character set to utf8 in the test on 5.7. Utf8mb4 is the default
character set in 8.0.1, so no changes to the test have to be made.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 16278