Bug #18301877	RPL.RPL_FLUSH_LOGS TEST FAILS OCASIONALLY ON PB2: FILE_EXISTS" ERROR 1
Analysis: In the test script, after executing "Flush logs" command
on Master, test script expects slave to create two relay log files.
Before checking for the existence of these files, test script is
doing "sync_slave_io_with_master.inc". This works fine if the test
script runs with 'GTID_MODE=OFF' because sync is checked with
positions. In case of GTID_MODE=ON , sync is checked with gtid
numbers. Flush logs (admin commands) does not generate a gtid number.
So sync_slave_io_with_master.inc (with gtid_mode ON mode) assumes
that slave is in sync with master even though it is not yet
executed latest "flush logs" command from the master.

Fix: "Flush" commands are logged as rotate events in the binary
log without gtid events. So it is correct to check master
and slave sync logic with positions instead gtids numbers.
Hence for this test script, making use_gtids to false
in all the cases (GTID_MODE ON or OFF).