BUG#26134361: GROUP_REPLICATION.GR_KEY_ROTATION TEST FAILING FREQUENTLY ON PB2 MYSQL-5.7
The test gr_key_rotation instruments some failures on the keyring
plugin, such as "keyring not found", which contains the full path.
The suppression of the error message was not considering correctly
Windows path separator character.

This patch fixes that by ignoring both '/' and '\' characters on the
error message suppression.