Bug#26138592 MANY DELETES (OR MANY INSERTS) THEN DROP TABLE MAY CAUSE UNDEFINED BEHAVIOUR
When a fragment is about to be released (releaseFragResources), make sure
that there are no pending expand or shrink on fragment.

As optimization change fragment state so that expand or shrink will do
nothing.

In addition to the above bug fix, in execEXPANDCHECK2 and execSHRINKCHECK2
check of fragment level is moved before potentially pages allocation.