WL#7194   Define and implement authorization model to manage XA-transactions
This patch introduces the new system privilege XA_RECOVER_ADMIN that controls
whether a user can execute the statement XA RECOVER in order to get
a list of XID for prepared XA transactions. The statement XA RECOVER can be
executed successfully only in case it is run by a user with the privilege
XA_RECOVER_ADMIN. It means that a user who is the initiator of the transaction
wouldn't be able to run XA RECOVER unless the user has the privilege
XA_RECOVER_ADMIN. By default the privilege XA_RECOVER_ADMIN isn't granted for
a new created user and explicit execution of the statement GRANT XA_RECOVER_ADMIN
is required before a user can run the statement XA RECOVER. In case the statement
XA RECOVER is failed by the reason of missing the privilege the error
ER_XAER_RMERR is issued. Additionally, the error ER_SPECIFIC_ACCESS_DENIED_ERROR
is issued as a warning to provide more information about a reason of statement
failure.