Bug#25473794: GR: PARTITIONED NODE SHOULD GET UPDATED STATUS AND NOT ACCEPT WRITES
When there is a network partition and a member is on a minority all
queries into that member will block.
In an effort to improve this situation, a new variable is added.
This variable will allow the user to chose if the member shall leave
the group after N seconds of being in a minority.