BUG#26121192 : --NO-CONSOLE IS COMPULSORY IF --LOG-ERROR USED TO MAKE                WINDOWS RUN PASS
Issue:
------
MTR adds '--console' option if '--no-console' options is not specified.
But adding '--console' option when '--log-error' is specified either in
test-[master/slave].opt or in the config file will cause the test run
to fail on windows.

Fix:
----
On windows, don't add '--console' option if '--log-error' is specified
either in test-[master/slave].opt or in the test config file. This patch
is a back-port of bug#20341933.

Reviewed-by: Deepa Dixit <deepa.dixit@oracle.com>
RB: 16372