Bug#25973525:INCORRECT WARNING IS LOGGED ON"FLUSH LOGS"WHEN BINARY-LOG IS IN USE AND ELD!=0
Problem:
With current behaviour if expire_logs_days/binlog_expire_log_seconds(on mysql-trunk)
is set, following warning will be logged whenever the user issues
"FLUSH LOGS"/"PURGE LOGS BEFORE" and binlog file is in use, though the file is
not old enough to get purged.
Eg:
2017-04-27T09:42:10.202909Z 10 [Warning] file ./master-bin.000001 was not purged
because it was being readby thread number 12

The warning is incorrect since it gives a notion to the user that system tried
to purge the log-file even when it was not old enough to get purged.

Analysis:
The root cause is that we check if a log is being used by other thread firstly
while purging it.

Fix:
To fix the problem, we check if a log older than the purge time firstly, then
check if it is being used by other thread while purging it. This will ensure
that the above warning is gone.

The changes in test rpl_4threads_deadlock was done because it started failing
after the changes done in binlog.cc. The reason for failure is explained below.

The sync point 'purge_logs_after_lock_index_before_thread_count' is set in the
method which checks if the log is in use, after the changes done in binlog.cc
this check(if the log is in use) will be done after purge_time check. The
purge_time in test was set as 9999-12-12 which was effectively being converted
to 0(Bug#26147576) and thus the sync point will never be reached.
To fix this the purge_time is modified to 2038-01-19(highest permissible value).

Additional changes done in rpl_4threads_deadlock
- We need to have the FLUSH LOGS command inside the while loop so that we have
  binary logs to flush in second iteration as well, as the PURGE LOGS command
  will purge all but the current binary log file.