WL#10611 Group Replication: Disallow writes after leave group
This worklog improves the safeness on operations around Group
Replication.

Currently when STOP GROUP_REPLICATION is executed the group replication stops
but clients are still able to do writes while the server is disconnected from
the group, thence not replicating the writes to the group.

This worklog enable super read only when STOP GROUP_REPLICATION is executed
preventing clients to execute writes after it.