Fix for bug#26106655 "DISCREPANCY BETWEEN IMPLICIT DB OF PARENT TABLE FOR FK AND REFERENCES PRIVILEGE".
Check for REFERENCES privilege was using incorrect database in some cases.

There was discrepancy between database which was used by SEs for parent
tables and used for check of REFERENCES privilege in cases when no explicit
database was specified for parent table.

This patch removes the discrepancy by aligning check for REFERENCES
privilege with SE behavior.