wl#10234, fix remaining review comments
Fixed an incorrect assert() which checked that the max
parrent correlation id was not exceeded.

Correlation Ids are produced starting from the specified
'batchRange' offset for a maximum of 'bs_rows'. Thus,
correlation Id's may be produced in the range
[batchRange .. batchRange + (bs_rows-1)].

The assert for not exceeding the MaxCorrelationId should
then be:

- Checked *before* sending the scanFragReq to the LQH node.
- Check for 'batchRange+bs_rows <= Max....' as corrIds are
  created starting with an offset of '0' and ending at
  an 'bs_rows-1' offset.