Punch hole support not available bug.
Issue
-----
Currently, some tests that require punch hole support to run are not
being run in some systems even though the punch hole support is
available for them.

The way we check if punch hole support is available or not is by running
a piece of c code which defines a macro without checking if it's defined
already or not. This missing check results in the compiler cribbing about
redefinition which the test misinterprets it as punch hole support not
available.

Fix
---
Add the #ifndef check to check for redefinition of the macro.

Reviewed-by: Jimmy Yang